# openapi.yaml
openapi: 3.1.0
info:
  title: Santa Monica Parcel Feasibility Engine API
  version: 0.1.0
  description: |
    AI-powered parcel evaluation and feasibility analysis tool for California housing development.

    This API analyzes parcels for development feasibility under California state housing laws including:
    - Base zoning regulations
    - SB9 lot split and duplex provisions
    - SB35 streamlined ministerial approval
    - AB2011 office-to-residential conversion
    - State Density Bonus Law
    - AB2097 parking reductions near transit

    The engine evaluates multiple development scenarios and recommends the optimal approach
    to maximize housing units while maintaining compliance with all applicable regulations.

    ## Authentication
    Currently no authentication required. Future versions will include API key authentication.

    ## Rate Limiting
    No rate limiting currently applied. Production deployment will implement rate limits.

    ## Errors
    The API uses standard HTTP status codes:
    - 200: Success
    - 400: Bad request (invalid parameters)
    - 404: Resource not found
    - 500: Server error
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.parcelfeasibility.com/api/v1
    description: Production server
...
paths:
...
  /rules/{rule_code}:
    get:
      tags:
        - Rules
      summary: Get details for specific rule
      description: |
        Returns detailed information about a specific California housing rule.

        Available rule codes:
        - SB9: Senate Bill 9 (lot splits and duplexes)
        - SB35: Senate Bill 35 (streamlined approval)
        - AB2011: Assembly Bill 2011 (office conversions)
        - AB2097: Assembly Bill 2097 (parking reductions)
        - DensityBonus: State Density Bonus Law
      operationId: getRuleDetails
      parameters:
        - name: rule_code
          in: path
          required: true
          description: Rule code (case insensitive)
          schema:
            type: string
            enum:
              - SB9
              - SB35
              - AB2011
              - AB2097
              - DensityBonus
          examples:
            sb9:
              value: "SB9"
              summary: Senate Bill 9
            sb35:
              value: "SB35"
              summary: Senate Bill 35

      responses:
        '200':
          description: Rule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleInfo'
              examples:
                sb9-details:
                  value:
                    name: "Senate Bill 9"
                    code: "SB9"
                    description: "Allows lot splits and up to two units per lot on parcels zoned for single-family"
                    year_enacted: 2021
                    applies_to:
                      - "Single-family residential zones"
                      - "Urban areas"
                    key_provisions:
                      - "Allows lot split creating two parcels minimum 1,200 sq ft each"
                      - "Allows up to 2 units per parcel (4 total with split)"
                      - "Ministerial approval required"
                      - "Maximum 800 sq ft per unit if lot < 5,000 sq ft"
                    eligibility_criteria:
                      - "Not in high fire hazard zone"
                      - "Not in flood zone"
                      - "Not historic property"
                      - "Not prime farmland"
                      - "Must be in urban area"

        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                not-found:
                  value:
                    detail: "Rule XYZ123 not found"
  /rules/check-eligibility/{rule_code}:
    get:
      tags:
        - Rules
      summary: Quick eligibility check for specific rule
      description: |
        Performs a quick eligibility check for a specific California housing rule
        without running a full analysis.

        Useful for:
        - Pre-screening parcels
        - Understanding basic requirements
        - Quick feasibility checks

        Note: This is a preliminary check. Full analysis is recommended for
        accurate development scenarios.
      operationId: checkRuleEligibility
      parameters:
        - name: rule_code
          in: path
          required: true
          description: Rule code to check eligibility
          schema:
            type: string
            enum:
              - SB9
              - SB35
              - AB2011
              - AB2097
              - DensityBonus

        - name: lot_size_sqft
          in: query
          required: true
          description: Lot size in square feet
          schema:
            type: number
            minimum: 0
          example: 6000

        - name: zoning_code
          in: query
          required: true
          description: Current zoning code
          schema:
            type: string
          example: "R1"

        - name: transit_distance_miles
          in: query
          required: false
          description: Distance to nearest major transit stop (for AB2097)
          schema:
            type: number
            minimum: 0
            default: 999.0
          example: 0.3

      responses:
        '200':
          description: Eligibility check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule_code:
                    type: string
                    description: Rule code checked
                  eligible:
                    type: boolean
                    description: Whether parcel is eligible
                  reasons:
                    type: array
                    items:
                      type: string
                    description: Reasons for eligibility determination
                  note:
                    type: string
                    description: Additional notes
              examples:
                sb9-eligible:
                  value:
                    rule_code: "SB9"
                    eligible: true
                    reasons:
                      - "Meets minimum lot size for SB9 lot split"
                    note: "This is a preliminary check. Full analysis recommended."

                sb9-not-eligible:
                  value:
                    rule_code: "SB9"
                    eligible: false
                    reasons:
                      - "Lot too small for SB9 lot split (need 2,400 sq ft minimum)"
                    note: "This is a preliminary check. Full analysis recommended."

                ab2097-eligible:
                  value:
                    rule_code: "AB2097"
                    eligible: true
                    reasons:
                      - "Within 0.5 miles of transit - no parking minimums required"
                    note: "This is a preliminary check. Full analysis recommended."

        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    RuleInfo:
      type: object
      required:
        - name
        - code
        - description
        - year_enacted
        - applies_to
        - key_provisions
        - eligibility_criteria
      properties:
        name:
          type: string
          description: Full name of the rule
          example: "Senate Bill 9"
        code:
          type: string
          description: Rule code or abbreviation
          example: "SB9"
        description:
          type: string
          description: Brief description of the rule
          example: "Allows lot splits and up to two units per lot on parcels zoned for single-family"
        year_enacted:
          type: integer
          description: Year the law was enacted
          example: 2021
        applies_to:
          type: array
          items:
            type: string
          description: Types of parcels or zones the rule applies to
          example:
            - "Single-family residential zones"
            - "Urban areas"
        key_provisions:
          type: array
          items:
            type: string
          description: Key provisions of the law
          example:
            - "Allows lot split creating two parcels minimum 1,200 sq ft each"
            - "Allows up to 2 units per parcel (4 total with split)"
        eligibility_criteria:
          type: array
          items:
            type: string
          description: Eligibility criteria for using this rule
          example:
            - "Not in high fire hazard zone"
            - "Not in flood zone"
            - "Not historic property"
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Analysis failed: Invalid parcel data"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)

# Future implementation
# security:
#   - ApiKeyAuth: []